requirements.txt
Flask==2.0.1
psycopg2-binary
#python-syslog==0.1.0
pytest==6.2.4
pytest-cov==2.12.1
python-dotenv
SQLAlchemy